requirements.txt
fastapi
uvicorn

sqlalchemy
psycopg2-binary

boto3

faiss-cpu
numpy

streamlit
requests